refactor: modularize summarize to lib
